+++ src/aks-preview/azext_aks_preview/_help.py
@@ -463,7 +463,7 @@ helps['aks create'] = """
         - name: --enable-keda
           type: bool
           short-summary: Enable KEDA workload auto-scaler.
-        - name: --enable-managed-cluster-snapshot
+        - name: --enable-cluster-snapshot
           type: bool
           short-summary: Cluster created can be snapshotable. 
         - name: --enable-defender